requirements.txt
fastapi[all]==0.75.2
aioredis==2.0.1
cos-python-sdk-v5==1.9.15
PyJWT==2.3.0
python-dotenv==0.19.1
tortoise-orm[aiomysql]==0.19.0
tencentcloud-sdk-python==3.0.618
qcloud-python-sts==3.1.1
wechatpy==2.0.0a11
qrcode[pil]==7.3.1
python-alipay-sdk==3.0.4
passlib==1.7.4
sqlmodel==0.0.6
fastapi-utils==0.2.1
fastapi-mail==1.1.0
docker==6.0.1
apscheduler
